harmonize flags & use epm's RunPackage
